pyproject.toml
[tool.poetry]
name = "readsync-ai"
version = "0.1.0"
description = ""
authors = ["Lei da sorte <75717568+Frederick528@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.13,<3.15"
fastapi = ">=0.128.0,<0.129.0"
uvicorn = {extras = ["standard"], version = ">=0.40.0,<0.41.0"}
pydantic-settings = ">=2.12.0,<3.0.0"
python-multipart = ">=0.0.21,<0.0.22"
huggingface-hub = {extras = ["inference"], version = ">=1.3.2,<2.0.0"}
python-dotenv = ">=1.2.1,<2.0.0"
numpy = ">=2.4.1,<3.0.0"
requests = ">=2.32.5,<3.0.0"
aioboto3 = ">=15.5.0,<16.0.0"
openai = ">=1.0.0"

[tool.poetry.group.dev.dependencies]
types-aioboto3 = {extras = ["s3"], version = ">=15.5.0,<16.0.0"}

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"   